+++ packages/fiber/src/core/utils/three-utils.ts
@@ -1,0 +1,65 @@
+import { Dpr, Size } from '../store'
+
+export type Camera = (THREE.OrthographicCamera | THREE.PerspectiveCamera) & { manual?: boolean }
+
+export const isOrthographicCamera = (def: Camera): def is THREE.OrthographicCamera =>
+  def && (def as THREE.OrthographicCamera).isOrthographicCamera
+
+export const isObject3D = (object: any): object is THREE.Object3D => object?.isObject3D
+
+export interface ObjectMap {
+  nodes: { [name: string]: THREE.Object3D }
+  materials: { [name: string]: THREE.Material }
+}
+
+export function calculateDpr(dpr: Dpr): number {
+  // Err on the side of progress by assuming 2x dpr if we can't detect it
+  // This will happen in workers where window is defined but dpr isn't.
+  const target = typeof window !== 'undefined' ? window.devicePixelRatio ?? 2 : 1
+  return Array.isArray(dpr) ? Math.min(Math.max(dpr[0], target), dpr[1]) : dpr
+}
+
+export function updateCamera(camera: Camera, size: Size): void {
+  // https://github.com/pmndrs/react-three-fiber/issues/92
+  // Do not mess with the camera if it belongs to the user
+  if (!camera.manual) {
+    if (isOrthographicCamera(camera)) {
+      camera.left = size.width / -2
+      camera.right = size.width / 2
+      camera.top = size.height / 2
+      camera.bottom = size.height / -2
+    } else {
+      camera.aspect = size.width / size.height
+    }
+    camera.updateProjectionMatrix()
+    // https://github.com/pmndrs/react-three-fiber/issues/178
+    // Update matrix world since the renderer is a frame late
+    camera.updateMatrixWorld()
+  }
+}
+
+// Collects nodes and materials from a THREE.Object3D
+export function buildGraph(object: THREE.Object3D): ObjectMap {
+  const data: ObjectMap = { nodes: {}, materials: {} }
+  if (object) {
+    object.traverse((obj: any) => {
+      if (obj.name) data.nodes[obj.name] = obj
+      if (obj.material && !data.materials[obj.material.name]) data.materials[obj.material.name] = obj.material
+    })
+  }
+  return data
+}
+
+export interface Disposable {
+  type?: string
+  dispose?: () => void
+}
+
+// Disposes an object and all its properties
+export function dispose<T extends Disposable>(obj: T): void {
+  if (obj.type !== 'Scene') obj.dispose?.()
+  for (const p in obj) {
+    const prop = obj[p] as Disposable | undefined
+    if (prop?.type !== 'Scene') prop?.dispose?.()
+  }
+}
