+++ packages/fiber/src/three-types.ts
@@ -1,9 +1,8 @@
 import type * as THREE from 'three'
 import type { EventHandlers, InstanceProps, ConstructorRepresentation } from './core'
-import type { Mutable, Overwrite } from './core/utils'
+import type { Mutable, Overwrite } from './core/utils/type-utils'
 
 export { Overwrite }
-
 interface MathRepresentation {
   set(...args: number[]): any
 }
